refactor and rename jobStorer and jobStore to dataStorer and dataStore respectively
