+++ api/tasks.go
@@ -14,7 +14,7 @@ import (
 var invalidBodyErrorMessage = "invalid request body"
 
 // CreateTaskHandler returns a function that generates a new TaskName resource containing default values in its fields.
-func (api *JobStoreAPI) CreateTaskHandler(ctx context.Context) http.HandlerFunc {
+func (api *DataStoreAPI) CreateTaskHandler(ctx context.Context) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
 		ctx := req.Context()
 		vars := mux.Vars(req)
@@ -34,7 +34,7 @@ func (api *JobStoreAPI) CreateTaskHandler(ctx context.Context) http.HandlerFunc 
 			return
 		}
 
-		newTask, err := api.jobStore.CreateTask(ctx, jobID, taskToCreate.TaskName, taskToCreate.NumberOfDocuments)
+		newTask, err := api.dataStore.CreateTask(ctx, jobID, taskToCreate.TaskName, taskToCreate.NumberOfDocuments)
 		if err != nil {
 			log.Error(ctx, "creating and storing a task failed", err, log.Data{"job id": jobID})
 			if err == mongo.ErrJobNotFound {
@@ -64,14 +64,14 @@ func (api *JobStoreAPI) CreateTaskHandler(ctx context.Context) http.HandlerFunc 
 }
 
 // GetTaskHandler returns a function that gets a specific task, associated with an existing Job resource, using the job id and task name passed in.
-func (api *JobStoreAPI) GetTaskHandler(w http.ResponseWriter, req *http.Request) {
+func (api *DataStoreAPI) GetTaskHandler(w http.ResponseWriter, req *http.Request) {
 	ctx := req.Context()
 	vars := mux.Vars(req)
 	id := vars["id"]
 	taskName := vars["task_name"]
 	logData := log.Data{"job_id": id, "task_name": taskName}
 
-	task, err := api.jobStore.GetTask(req.Context(), id, taskName)
+	task, err := api.dataStore.GetTask(ctx, id, taskName)
 	if err != nil {
 		log.Error(ctx, "getting task failed", err, logData)
 		if err == mongo.ErrJobNotFound {
